+++ packages/core/src/mixins/closable.js
@@ -5,7 +5,7 @@ import keycode from 'keycode';
 import getComponentName from '../utils/getComponentName';
 
 const ESCAPE = 'Escape';
-const TOUCH_CLOSE_DELAY = 200;
+const TOUCH_CLOSE_DELAY = 100;
 
 /**
  * `closable()` HOC mixin
@@ -37,18 +37,19 @@ const closable = ({
 
         state = {
             closeDelayTimeout: null,
+            clickedInside: false,
         };
 
         componentDidMount() {
             document.addEventListener('keyup', this.handleDocumentKeyup);
-            document.addEventListener('click', this.handleDocumentClick);
-            document.addEventListener('touchstart', this.handleDocumentTouch);
+            document.addEventListener('click', this.handleDocumentClickOrTouch);
+            document.addEventListener('touchstart', this.handleDocumentClickOrTouch);
         }
 
         componentWillUnmount() {
             document.removeEventListener('keyup', this.handleDocumentKeyup);
-            document.removeEventListener('click', this.handleDocumentClick);
-            document.removeEventListener('touchstart', this.handleDocumentTouch);
+            document.removeEventListener('click', this.handleDocumentClickOrTouch);
+            document.removeEventListener('touchstart', this.handleDocumentClickOrTouch);
             clearTimeout(this.state.closeDelayTimeout);
         }
 
@@ -62,11 +63,8 @@ const closable = ({
         }
 
         /**
-         * Touchstart event fires as soon as finger touches the screen,
-         * emulate ghost-click's time delay to fire child events first
-         * before trigger the onClose event.
-         *
-         * @ref http://stackoverflow.com/a/9634715
+         * Delay slightly to fire child events first
+         * before trigger the `onClose` event.
          */
         delayedClose = (event) => {
             const timeout = setTimeout(
@@ -78,45 +76,35 @@ const closable = ({
 
         captureInsideEvents = (node) => {
             if (node) {
-                node.addEventListener('click', this.handleInsideClick);
-                node.addEventListener('touchstart', this.handleInsideTouch);
+                node.addEventListener('click', this.handleInsideClickOrTouch);
+                node.addEventListener('touchstart', this.handleInsideClickOrTouch);
             }
             this.nodeRef = node;
         }
 
-        // Trigger `onClose()` call on Escape keyup if turned on in options.
         handleDocumentKeyup = (event) => {
             if (onEscape && event.keyCode === keycode(ESCAPE)) {
                 this.props.onClose(event);
             }
         }
 
-        handleDocumentClick = () => {
-            if (onClickOutside) {
-                this.props.onClose();
+        handleDocumentClickOrTouch = (event) => {
+            if (this.state.clickedInside) {
+                this.setState({ clickedInside: false });
+                return;
             }
-        }
 
-        // Trigger `onClose()` call on any touch if turned on in options.
-        handleDocumentTouch = (event) => {
             if (onClickOutside) {
                 this.delayedClose(event);
             }
         }
 
-        handleInsideClick = (event) => {
-            if (onClickInside) {
-                this.props.onClose(event);
-            }
-            // Stop event from bubbling up so the listeners at document won't hear.
-            event.stopPropagation();
-        }
+        handleInsideClickOrTouch = (event) => {
+            this.setState({ clickedInside: true });
 
-        handleInsideTouch = (event) => {
             if (onClickInside) {
                 this.delayedClose(event);
             }
-            event.stopPropagation();
         }
 
         render() {