+++ packages/storybook/examples/Popover/DemoList.js
@@ -1,20 +1,31 @@
 import React from 'react';
+import { action } from '@storybook/addon-actions';
 
+import Button from '@ichef/gypcrete/src/Button';
 import List from '@ichef/gypcrete/src/List';
 import ListRow from '@ichef/gypcrete/src/ListRow';
-import TextLabel from '@ichef/gypcrete/src/TextLabel';
+
+function DemoButton(props) {
+    return (
+        <Button
+            bold
+            minified={false}
+            color="black"
+            {...props} />
+    );
+}
 
 function DemoList() {
     return (
         <List>
             <ListRow>
-                <TextLabel basic="Row 1" />
+                <DemoButton basic="Row 1" onClick={action('click.1')} />
             </ListRow>
             <ListRow>
-                <TextLabel basic="Row 2" />
+                <DemoButton basic="Row 2" onClick={action('click.2')} />
             </ListRow>
             <ListRow>
-                <TextLabel basic="Row 3" />
+                <DemoButton basic="Row 3" onClick={action('click.3')} />
             </ListRow>
         </List>
     );